Update main.js
Added function called changeColor(color) that will update the stroke style. It is possible to just enter a string instead of using hexadecimal (use "green" or "black"). 

Another thing we should add is having an initializing function that will initialize everything once the document has loaded. This can be done with adding an init() function that is an event listener of document.onload.